Add trigger property to ThreadResopnse
