+++ programs/network/src/jobs/delete_snapshot/job.rs
@@ -38,5 +38,6 @@ pub fn handler(ctx: Context<DeleteSnapshotJob>) -> Result<ThreadResponse> {
             ],
             data: anchor_sighash("delete_snapshot_process_snapshot").to_vec(),
         }),
+        trigger: None,
     })
 }
