+++ programs/network/src/jobs/distribute_fees/job.rs
@@ -41,5 +41,6 @@ pub fn handler(ctx: Context<DistributeFeesJob>) -> Result<ThreadResponse> {
             ],
             data: anchor_sighash("distribute_fees_process_snapshot").to_vec(),
         }),
+        trigger: None,
     })
 }
