+++ programs/network/src/jobs/distribute_fees/process_snapshot.rs
@@ -64,5 +64,6 @@ pub fn handler(ctx: Context<DistributeFeesProcessSnapshot>) -> Result<ThreadResp
         } else {
             None
         },
+        trigger: None,
     })
 }
