+++ programs/network/src/jobs/process_unstakes/unstake_process.rs
@@ -152,5 +152,8 @@ pub fn handler(ctx: Context<UnstakeProcess>) -> Result<ThreadResponse> {
         None
     };
 
-    Ok(ThreadResponse { next_instruction })
+    Ok(ThreadResponse {
+        next_instruction,
+        trigger: None,
+    })
 }
