+++ programs/network/src/jobs/stake_delegations/process_delegation.rs
@@ -139,5 +139,8 @@ pub fn handler(ctx: Context<StakeDelegationsProcessDelegation>) -> Result<Thread
         None
     };
 
-    Ok(ThreadResponse { next_instruction })
+    Ok(ThreadResponse {
+        next_instruction,
+        trigger: None,
+    })
 }
