+++ programs/network/src/jobs/stake_delegations/process_worker.rs
@@ -78,5 +78,8 @@ pub fn handler(ctx: Context<StakeDelegationsProcessWorker>) -> Result<ThreadResp
         None
     };
 
-    Ok(ThreadResponse { next_instruction })
+    Ok(ThreadResponse {
+        next_instruction,
+        trigger: None,
+    })
 }
