+++ programs/network/src/jobs/take_snapshot/create_snapshot.rs
@@ -76,5 +76,6 @@ pub fn handler(ctx: Context<TakeSnapshotCreateSnapshot>) -> Result<ThreadRespons
         } else {
             None
         },
+        trigger: None,
     })
 }
