+++ programs/network/src/jobs/take_snapshot/job.rs
@@ -41,5 +41,6 @@ pub fn handler(ctx: Context<TakeSnapshotJob>) -> Result<ThreadResponse> {
             ],
             data: anchor_sighash("take_snapshot_create_snapshot").to_vec(),
         }),
+        trigger: None,
     })
 }
