+++ programs/thread/src/instructions/thread_create.rs
@@ -5,6 +5,7 @@ use    anchor_lang::{
         solana_program::system_program,
         system_program::{transfer, Transfer},
     };
+use clockwork_utils::InstructionData;
 
 use crate::state::*;
