+++ programs/thread/src/state/mod.rs
@@ -1,7 +1,6 @@
 //! All objects needed to describe and manage the program's state.
 
 mod thread;
-mod utils;
 
+pub use clockwork_utils::*;
 pub use thread::*;
-pub use utils::*;
